rebase and restructure tests
