+++ packages/web-app-files/tests/unit/helpers/resource/common.spec.ts
@@ -1,4 +1,4 @@
-import { buildQueryString } from '../../../src/helpers/resource/common'
+import { buildQueryString } from '../../../../src/helpers/resource/common'
 
 describe('common', () => {
   test('buildQueryString', () => {
