+++ packages/web-app-files/tests/unit/helpers/resource/privatePreviewBlob.spec.ts
@@ -1,4 +1,4 @@
-import { privatePreviewBlob } from '../../../src/helpers/resource'
+import { privatePreviewBlob } from '../../../../src/helpers/resource'
 import mockAxios from 'jest-mock-axios'
 import merge from 'lodash-es/merge'
