+++ packages/web-app-files/tests/unit/services/cache.spec.ts
@@ -1,4 +1,4 @@
-import { cacheService } from '../../src/services'
+import { cacheService } from '../../../src/services'
 import { Cache } from 'web-pkg/src/cache'
 
 describe('cache', () => {
