+++ packages/web-app-files/tests/unit/services/client.spec.ts
@@ -1,4 +1,4 @@
-import { clientService } from '../../src/services'
+import { clientService } from '../../../src/services'
 import { HttpClient } from 'web-pkg/src/http'
 import mockAxios from 'jest-mock-axios'
