+++ tests/integration/config/jest.init.js
@@ -5,7 +5,13 @@ import ODS from 'owncloud-design-system'
 
 // eslint-disable-next-line jest/no-mocks-import
 import sdkMock from '../../../__mocks__/sdk'
-import encodePath from '@runtime/src/helpers/encodePath'
+import { encodePath } from '@pkg/src/utils'
+
+window.IntersectionObserver = jest.fn(() => ({
+  observe: jest.fn(),
+  disconnect: jest.fn(),
+  unobserve: jest.fn()
+}))
 
 Vue.use(ODS)
